LPS-196123 Move class DLPreviewableProcessor and its subclasses into module
# breaking_change_report

## What `portal-impl/src/com/liferay/portlet/documentlibrary/util/DLPreviewableProcessor.java`, `portal-impl/src/com/liferay/portlet/documentlibrary/util/AudioProcessorImpl.java` and `portal-impl/src/com/liferay/portlet/documentlibrary/util/VideoProcessorImpl.java`

`DLPreviewableProcessor` is moved from `portal-impl` into `document-library-preview-api`, `AudioProcessorImpl` is moved from `portal-impl` into `document-library-preview-audio` and `VideoProcessorImpl` is moved from `portal-impl` into `document-library-preview-video`.

## Why

We want to make `AudioProcessorImpl` and `VideoProcessorImpl` components of OSGI, all subclasses of `DLPreviewableProcessor` will be in the module, so there is no need for it to be in `portal-impl`.

## Alternatives

Add `document-library-preview-api`, `document-library-preview-audio` or `document-library-preview-video` in build dependency and use the same class.

----